requirements.txt
python-barcode>=0.15
Pillow>=9.0
pyserial>=3.5
